requirements.txt
contourpy==1.0.7
cycler==0.11.0
ecal5==5.10.1
fonttools==4.39.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pillow==9.4.0
protobuf==3.11.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
